备份

+++ src/org/itxtech/synapseapi/runnable/SendChangeDimensionRunnable.java
@@ -1,6 +1,9 @@
 package org.itxtech.synapseapi.runnable;
 
+import cn.nukkit.Player;
+import cn.nukkit.Server;
 import cn.nukkit.network.protocol.ChangeDimensionPacket;
+import cn.nukkit.network.protocol.DataPacket;
 import cn.nukkit.network.protocol.PlayStatusPacket;
 import org.itxtech.synapseapi.SynapsePlayer;
 
@@ -20,7 +23,7 @@ public class SendChangeDimensionRunnable implements Runnable {
     @Override
     public void run() {
         ChangeDimensionPacket changeDimensionPacket1 = new ChangeDimensionPacket();
-        changeDimensionPacket1.dimension = (byte) this.dimension;
+        changeDimensionPacket1.dimension = this.dimension;
         changeDimensionPacket1.x = (float) player.getX();
         changeDimensionPacket1.y = (float) player.getY();
         changeDimensionPacket1.z = (float) player.getZ();
